requirements.txt
opentelemetry-api==1.11.0
opentelemetry-sdk==1.11.0
opentelemetry-instrumentation==1.11.0
opentelemetry-exporter-otlp==1.11.0
